Rename python module backend to algo
