+++ .coveragerc
@@ -2,7 +2,7 @@
 plugins = Cython.Coverage
 include = dpnp/*,examples/*,build_cython/dpnp/*
 branch = True
-source = dpnp,dpnp.backend
+source = dpnp,dpnp.dpnp_algo
 
 [report]
 exclude_lines =
