+++ dpnp/dpnp_algo_indexing.pyx
