+++ dpnp/dpnp_algo_logic.pyx
