+++ dpnp/dpnp_algo_manipulation.pyx
