+++ dpnp/dpnp_algo_mathematical.pyx
