+++ dpnp/dpnp_algo_searching.pyx
