+++ dpnp/dpnp_algo_trigonometric.pyx
