+++ dpnp/dpnp_iface.py
@@ -44,7 +44,7 @@ import os
 import numpy
 import collections
 
-from dpnp.backend import *
+from dpnp.dpnp_algo import *
 from dpnp.dparray import dparray
 from dpnp.dpnp_utils import *
 from dpnp.fft import *
