+++ dpnp/dpnp_iface_searching.py
@@ -42,7 +42,7 @@ it contains:
 
 import numpy
 
-from dpnp.backend import *
+from dpnp.dpnp_algo import *
 from dpnp.dparray import dparray
 from dpnp.dpnp_utils import *
